refactor(tests): address PR review feedback
- Unexport WorkerNodeScope, KVMAllSchedulableNodeScope,
  KVMFallbackDiagnostic — only used within the vmhelpers package.
- Remove backward-compat shims in vmhelpers/log_matchers.go and
  vmhelpers/k8s_client.go; move their tests to logmatchers/ and
  k8sutil/ respectively, since those packages are new and have no
  external consumers.
- Use defer for context cancellation in VerifyClusterVSOCKReadyPhases
  to prevent context leak on panic.

Partially generated by AI.
